backend/requirements.txt
# Web 框架
Flask==3.0.3
flask-cors==4.0.0

# 数据库
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
PyMySQL==1.1.1
cryptography>=42.0.0

# 认证
PyJWT==2.8.0

# 图像处理
Pillow==10.3.0

# AI 推理
# 建议优先通过官方 CPU 源安装：
# pip install torch torchvision --index-url https://download.pytorch.org/whl/cpu
# requirements 中保留最低版本要求，交由 pip 根据当前 Python 版本选择兼容 wheel
torch>=2.3.0
torchvision>=0.18.0

# 缓存
redis==5.0.4

# 环境变量
python-dotenv==1.0.1

# 语音识别（百度）
baidu-aip==4.16.13

# WSGI 服务
gunicorn==22.0.0

# 测试
pytest==8.2.2
pytest-flask==1.3.0

# HTTP 客户端
requests==2.32.3
